Move @veto/db FE files to veto-app/frontend Move @veto/db BE files to veto-app/backend Add fix typing issues with gunClient & store Ensure @veto/db is built before starting the application Ensure @veto/db out/ directory is cleaned before building
