+++ package.json
@@ -15,7 +15,7 @@
     "packages/shared/*"
   ],
   "scripts": {
-    "prebuild": "yarn workspace @veto/config build && yarn workspace @maintenance-app/types build",
+    "prebuild": "yarn workspace @veto/config build && yarn workspace @veto/db build && yarn workspace @maintenance-app/types build",
     "build": "yarn workspace veto-backend build && yarn build:veto:frontend && yarn workspace screen-app build",
     "build:veto:frontend": "yarn workspace veto-frontend build",
     "cyphernode:setup": "SKIP_VETO=true ./scripts/setup.sh",
@@ -24,7 +24,7 @@
     "dev": "yarn cyphernode:start && yarn start",
     "lint": "prettier -l '**/*.{ts,tsx,js,jsx,json}'",
     "lint:fix": "prettier --write '**/*.{ts,tsx,js,jsx,json}'",
-    "prestart": "yarn workspace @veto/config build && yarn workspace @maintenance-app/types build",
+    "prestart": "yarn prebuild",
     "start": "yarn start:veto",
     "start:backend": "yarn start:veto:backend",
     "start:frontend": "yarn start:veto:frontend",
