+++ packages/veto-app/shared/db/src/index.ts
@@ -1,13 +1,11 @@
-const gun = require('./gun')
-
-export { default as StoreSettings, Settings } from './settings'
-export { default as StoreNotifications, Notification } from './notifications'
-export {
-  default as StoreTransactions,
-  TransactionType,
-  TransactionStatus,
-  TransactionHistory,
-  Transaction,
-} from './transactions'
+//export { default as StoreSettings, Settings } from './settings'
+//export { default as StoreNotifications, Notification } from './notifications'
+//export {
+//  default as StoreTransactions,
+//  TransactionType,
+//  TransactionStatus,
+//  TransactionHistory,
+//  Transaction,
+//} from './transactions'
 
 export default gun
